fixed weird error with "Nesting::vector<uint32_t> parents"
for some reason Nesting::vector<uint32_t> parents gives a fatal error sometimes when trying to delete Nesting object

+++ soul_cppTranspiller/soul_cppTranspiller/tokenizer.cpp
@@ -44,7 +44,7 @@ struct CharIterator
 
     bool setOffset(int64_t offset)
     {
-        if (index + offset < 0 || index + offset > str.size())
+        if (index + offset < 0 || index + offset > (int64_t)str.size())
             return false;
 
         index = index + offset;
